requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
cryptography==39.0.0
distlib==0.3.6
Django==4.1.5
django-4-jet==1.0.9
django-cors-headers==3.13.0
django-cors-middleware==1.5.0
django-jet==1.0.8
djangorestframework==3.14.0
filelock==3.9.0
mysqlclient==2.1.1
Pillow==9.4.0
pipenv==2022.12.19
platformdirs==2.6.2
pycparser==2.21
Pygments==2.14.0
PyMySQL==1.0.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
virtualenv==20.17.1
virtualenv-clone==0.5.7
whitenoise==6.3.0
